changed model, some songs, updated visualizer and training to use 8 bit

+++ WaveNET/ops.py
@@ -81,7 +81,7 @@ def mu_law_decode(output, quantization_channels):
     with tf.name_scope('decode'):
         mu = quantization_channels - 1
         # Map values back to [-1, 1].
-        signal = 2 * (tf.cast(output, tf.float32) / mu) - 1 # no gradient?? tf.to_float(output)
+        signal = 2 * (tf.to_float(output) / mu) - 1 # no gradient?? tf.to_float(output)
         # Perform inverse of mu-law transformation.
         magnitude = (1 / mu) * ((1 + mu)**abs(signal) - 1)
         return tf.sign(signal) * magnitude
